load and store panels
